requirements_instagram_gemini.txt
# Requirements for Instagram Gemini Scraper
instaloader>=4.9.5
google-generativeai>=0.3.0
python-dotenv>=1.0.0